added content actions
